Implement registration (#5)
As part of this work, the HTTP client's error handling was made a bit
more consistent. When the API returns a 4xx or 5xx status, the body is
always a JSON object with an "errors" key holding an array of strings.

There is a new PlateZeroApiError class which extends Error and has a
`messages` property, which allows us to easily handle the multiple
errors that may be returned from an API call.

As needed, we can add to the errors.ts file to support additional error
types.