Enabled editing of existing ingredient lists
Renamed NewIngredientList to simply IngredientList and added an
`ingredientList` prop so that if we have an existing ingredient list,
this component could be used to display/edit it.

This required reconciling some differences between the IngredientList
model and IngredientListJSON: in the IngredientListJSON, the `lines`
property was called `ingredients`. Now, it is called `lines` to match
the database model.

Additionally, the filename for the NewRecipeTitle component has been
changed from NewRecipe to match its component name.