Use a hook to provide cleaner access to auth state
React has a new Hook feature which enables function components
(previously called "stateless") to use some elements of state. This
often results in cleaner code, and can also provide slightly easier
access to context.

This is a demo/proof of concept for refactoring the PlateZeroContext to
simply be a UserContext to access the current user, and shows how it can
be used by refactoring the Navbar component to be a functional rather
than class-based component.