Change "Servings" to "Yield"
We store the recipe's yield as a generic text string rather than a
number of servings as it allows us to be more specific about the output
of a recipe. For example, you might have an ice cream that yields "1
pint", or a cookie recipe that yields "24 large or 36 small cookies".