Enable editing of multiple ingredient lists
Move away from separating out added lists/removed IDs and simply include
the effective list of procedure/ingredient lists/lines in the
RecipeVersionPatch. This simplifies things at several levels, including
the ability to consolidate logic for handling updating/removal/addition
of items in a list in changes.ts. It also vastly simplifies the process
of applying the patch from the server's perspective. Additionally, this
paves the way for support for drag-and-drop reordering of
steps/ingredients.

When applying a patch, the server checks to see whether the item has an
undefined `id` field. If so, the interpretation is to create a new
procedure list/procedure line/ingredient list/ingredient line. If the
`id` field is NOT undefined, then the other fields are ignored and the
existing row is referenced. This means that updates are tracked the same
way that additions are; by "editing" e.g. a procedure step, the
effective operation is to replace it in the procedure list with a new
step. In the future, the edit history can be reconstructed by noting the
differing row IDs of the items.