Show a single line of recipes on user page unless expanded
To make it more visually apparent that the recipes displayed on the user
profile page is a limited subset, only display a single row of recipes,
responsive to the size of the screen. The final card on the line becomes
a "see more" card when there are more recipes to display than will fit
on a single line.

The exception to this is the smallest responsive screen size, where each
recipe card takes up an entire line. In this case, we display a maximum
of three recipes before showing the "see more" card.

Fixes #58