Use shortid for UITrackables
This removes reliance on the database id in the JSON field and as a
result makes it possible to use the remove/replace/restore operations
for both lists and for objects.

In the future, we may want to replace the sequential IDs for recipe
versions, procedure lists, et al, with shortids as the primary key.
The concept of a sequence often doesn't make sense (e.g. the first
version of a recipe might have an id of 600); the more analogous concept
is a _random_ unique value. This would slightly simplify the tracking of
changes as there would be only one `id` which could be generated on the
client side for new items. Furthermore, it would make offline editing
much more attainable.