Clean up UI for single procedure/ingredient lists
Since most recipes only have a single ingredient and procedure list,
this default case should be as concise as possible. This means we don't
need to spend a bunch of space on the left margin to manage multiple
procedure/ingredient lists until the user has actually added additional
lists that they need to manage.