Omit ID when converting to JSON, not when replacing
If the ID is omitted when an item is changed, a new row will always be
created in the database. By moving the id omission to when we are
actually converting the UITrackable back into a JSON object, we can
check to see whether there is an effective change between the current
value and the original, and omit the ID here if that is the case.