Support for multiple procedure/ingredient lists (#65)
* Use Row/Col for edit recipe header

* Make ingredient line inputs more responsive

* Hide some chrome when editing

Editing is quickly becoming very busy; this should help calm it down a
bit.

* Enable editing of multiple ingredient lists

Move away from separating out added lists/removed IDs and simply include
the effective list of procedure/ingredient lists/lines in the
RecipeVersionPatch. This simplifies things at several levels, including
the ability to consolidate logic for handling updating/removal/addition
of items in a list in changes.ts. It also vastly simplifies the process
of applying the patch from the server's perspective. Additionally, this
paves the way for support for drag-and-drop reordering of
steps/ingredients.

When applying a patch, the server checks to see whether the item has an
undefined `id` field. If so, the interpretation is to create a new
procedure list/procedure line/ingredient list/ingredient line. If the
`id` field is NOT undefined, then the other fields are ignored and the
existing row is referenced. This means that updates are tracked the same
way that additions are; by "editing" e.g. a procedure step, the
effective operation is to replace it in the procedure list with a new
step. In the future, the edit history can be reconstructed by noting the
differing row IDs of the items.

* Use shortid for UITrackables

This removes reliance on the database id in the JSON field and as a
result makes it possible to use the remove/replace/restore operations
for both lists and for objects.

In the future, we may want to replace the sequential IDs for recipe
versions, procedure lists, et al, with shortids as the primary key.
The concept of a sequence often doesn't make sense (e.g. the first
version of a recipe might have an id of 600); the more analogous concept
is a _random_ unique value. This would slightly simplify the tracking of
changes as there would be only one `id` which could be generated on the
client side for new items. Furthermore, it would make offline editing
much more attainable.

* Use refactored tracking logic for ingredient lists

* Fix new recipe creation