Add a bottom border to plain inputs
Currently, it's a little hard to tell that these are actually input
fields. This patch adds a bottom border to the PlainInput and
PlainSelect components which should help with this. An issue where the
background color of PlainSelect is being overridden to white resulting
in improper rendering of `bg-added` and `bg-changed` classes has been
fixed.

Additionally, the usage of PlainInput in the Preheats component has been
updated to be more consistent with its usage in the procedure/ingredient
sections.