Reorganize recipe layout
Display the recipe's vital statistics (duration, preheats, yield,
author, source, etc) in an inline list near the top with the description
below and the image, if any, before the ingredients.

This addresses a few cases that looked a bit weird previously, such as
when a recipe had only an image and no description, and the image would
be smashed over to the right hand side with an empty void to its left.

Since the source URL (if present) is now available in the vital
statistics, we no longer need to show a reference to it in the Actions
menu, which is also nice because it's not really an action in the same
sense that editing the recipe is.