Stretch "see more" link
Previously, while the recipe card had stretched links, the "see more"
link did not. A "stretched link" means that you can click anywhere on
the surrounding container to follow the link. Now, you can click
anywhere on the gray box to see additional recipes just as you can click
anywhere on the title or image of a recipe to navigate to it.