Show the recipe version being displayed
When we render a recipe version, show an unobtrusive summary of who
created the version and when, along with navigation to the previous
version if applicable. In the future, it would also be nice to provide a
diff view to easily see what changed between two versions, along with
"next" navigation (this gets slightly trickier than previous because
next is 1:n where prev is 1:1).