Make user and recipe list pages more distinct
The user detail page and the recipe list page were previously very
similar, but I think they serve very different purposes. The user detail
page is nice for being able to see who a person is and what they've been
up to in the kitchen. On the recipe page, I need a very concise way to
to look through potentially hundreds of recipes easily.

Previously, on the recipe list page, we were devoting way too much space
to each individual recipe making it pretty unusable on small screen
sizes in particular. I'm certain we'll want to tweak this more after
playing with it in prod, but I think it's at least a step in the right
direction.