Enable sign in with password manager
Password managers (or at least 1Password and Pass on iOS Safari) don't
send the `change` event when they fill in the login form. As a result,
React isn't made aware of the username and password being provided and
thinks that they are still both empty strings.

To resolve this, this patch uses a slightly more primitive approach
where, when the form is submitted, we simply read the value of the
inputs and submit those values to the endpoint. Unfortunately, this
means we can't effectively conditionally disable the login button, but I
think this is a pretty acceptable compromise for being able to use a
password manager.