Fix editing background colors, add back placeholders
Add back the placeholders even when not being hovered, now that the
borders are a little less aggressive.

Fix the added/changed background colors for the inputs (there were a few
weird behaviors such as where a changed line input was focused, the
background would go back to white).

Reintroduce the ability to reset a changed ingredient/procedure line
back to its original state by updating the `act` function to consider
whether the line has been changed, as well as setting the line icon
based on whether the line has been changed.