Fix amount parsing
This patch addresses two issues with amount parsing. First, the
ingredient

  1 15-ounce can diced tomatoes

was previously not being handled correctly (the "15" was considered part
of the quantity and the fraction library doesn't know what to do with
the string "1 15"). Furthermore, when a quantity was not being found, we
were falling back to using "1" as the numerator/denominator. This leads
to unexpected results, e.g. when the recipe calls for "kosher salt", we
shouldn't translate this to "1 kosher salt".

To address these items, I've refactored the regex to be a bit more
specific in what it's looking for and introduced a matchNumber function
which separates this logic out a bit from the main ingredient parsing
function.

Furthermore, I've updated/corrected the unit tests to check for the
updated amount parsing logic.

Fixes #88