Remove username from navbar
Since the username is a variable length string, it can easily lead to
weird overflow/wrapping issues when displayed along with the other items
in the navbar on small screens. One solution is to use the username in
the navbar and only display the current user's avatar to link them to
their profile page.

Fixes #86