Enable editing of yield and duration
Add yield and duration inputs to the new and edit recipe pages. For
simplicity, the duration is simply tracked as a number of hours and
minutes. I don't think anyone will really want to be more specific than
minutes, and I haven't seen a recipe that takes so long you'd want to
express it in units of time larger than hours.

This also fixes a bug with the calculation of hours and minutes as shown
in the recipe vitals; previously, the minutes being displayed were not
being subtracted from the total duration when figuring the hours. Now,
this logic is centralized in common/time.ts and used both by the recipe
vitals component and by the duration input component.

Finally, the .bg-changed and .bg-added global CSS classes have been
updated to be !important. This is because Bootstrap sets the
background-color of .form-control:focus to be white, which is
undesirable when the .bg-changed or .bg-added class is present.

Fixes #94