Show additional and improved hints about commit message
For most people who will probably be using PlateZero, a commit message
is a foreign concept. Eventually, we'll probably want some documentation
about what this message is, why it's helpful, and how to write one well.
But for now, we can improve the hints provided about what to do here.

This change adds a reminder that the change description is required as
well as adjusts the placeholder text to provide an example of what the
message could be.

Fixes #99