Implement basic history browsing
Modify the recipe page to show a tab bar with the default tab "Recipe"
(which shows the current recipe revision, as before) and a new "History"
tab which shows a basic timeline of recipe versions.

The Actions menu has been moved into the toolbar, and the recipe
description (if present) has been moved up above the toolbar to more
logically separate the Recipe from the RecipeVersion.

Now that we have a slightly more powerful means of browsing history, the
Previous Version link in the Authored By line has been moved into the
details disclosure.