Refresh Tokens (#106)
New table holding `refresh_tokens` for a user. On login and register, a
refresh token is also generated and given to the user. When a 401 is
returned from the server, and the client has a refresh token and regular
token, the client refreshes the JWT token and tries again.

This required some refactoring throughout the entire application.
Removed TokenContext and moved authentication management to the api
client, which is now stateful.