Show enhanced recipe source information
Now that we've added some additional recipe source fields, it would be
nice to display them on the recipe page. This patch adds support for
rendering the additional source fields based on which ones are present.
I think this should do a pretty good job for just about all of the
permutations as well as not break horribly for any of them. As we go, we
may discover a few improvements we can make, however.