Hide User#email in JSON
Currently, someone could browse to /api/users and get a list of all of
the email addresses that have been used to sign up for PZ. This isn't
great. We aren't really using the email addresses anywhere, so for now
we'll just hide them the same way we do the password hashes. I'm sure
once we start building out a profile page sort of thing, we'll need to
revisit this so that users can see/update their email address. I think
there may be some interesting things we can do with sequelize Scopes,
though I haven't dug into the docs for that quite enough to know for
sure, much less what exactly we'd want to do.