Add support for text child of recipeInstructions
This is the markup used on PlateZero (as a schema.org/HowToStep
itemtype). We're not capturing every single detail (e.g. step images,
etc) but this at least doesn't entirely skip the procedure when
importing from PZ to PZ. Eventually, what we'll really want is true
forking capability where this just happens in the database rather than
through the importer, but I think adding this for now lets us punt on
that for a bit longer.