Config Cleanup & Local Cookie Fix (#122)
`secure` tag on the authentication cookie meant it would not be saved
locally, where we don't use https. Fixed by setting the cookie
configuration in the Next.js build environment variables, which are
inserted during build time. Sets `secure` depending on the environment.

Also cleaned up the configuration management on the server side, just
using an object instead of a function.