Create an EditableImage component
There are several places where people want to be able to upload pictures
to include on the PZ website. The most immediate case is when a recipe
is imported from some website but the user wants to use a different
picture, or just add a picture when there previously was none.

The new EditableImage component provides a flexible mechanism to enable
user-provided images to be used on PlateZero. The API server has a new
endpoint, POST /api/user/images, where a user can upload arbitrary image
files and later access them through static.platezero.com. The
EditableImage component is designed to show an image along with a button
which launches a modal where a new image can be uploaded to the new
endpoint and previewed in the modal. When the user is satisfied with
their upload, they can click the "save" button in the modal which calls
a callback function from the parent component with the URL of the newly
uploaded image, allowing the parent component to use the new image
however is applicable.