Centralize HTTP response error rendering
In a lot of places, we wanted to get an array of error messages, either
by looking at the "errors" fields from JSON returned from the API, or by
grabbing the message property from a JavaScript Error object if it
wasn't an HTTP error.

Usually, we ended up showing this list of errors in a Bootstrap <Alert>
component. Rather than implementing the type checking everywhere, a new
helper function in common/http can be called with any error object in
order to get an array of error messages from whatever source is
appropriate.

Also, a new AlertErrors component can be used to conditionally render a
Bootstrap danger alert with a <ul> of the error messages inside it.