Use semicolon for ingredient prep separation
As it turns out, it can be hard to type two consecutive hyphens. Some
operating systems/browsers, including iOS, will automatically change --
into an em dash, which makes it very hard if not impossible to edit an
ingredient list.

Instead of using

    3 cloves of garlic -- minced

we will now use the format

    3 cloves of garlic; minced

which may actually be a little nicer to type (fewer characters) and
easier to read anyway.

Fixes #180