Handle toggle() called after unmount
When the user mouses off the tooltip target, there is a short delay
before toggle() is called to hide the tooltip. In the interim, it's
possible that the component is unmounted, e.g. during navigation to
another page.

As a hack around this, we set isMounted to true, and as a cleanup
effect, set isMounted to false. Then when toggle gets called, we can
check whether we should actually be mutating the state based on whether
the component is still mounted.