Move ingredient parsing into common library
Before, we had a bunch of ingredient parsing in different places, some
of it duplicated. This change introduces the use of a new, standalone,
open-source package "ingredient-parser" which lives at
https://github.com/plate0/ingredient-parser.

All unit tests and helper functions surrounding the parsing of a loosely
structured "ingredient" have been moved into the ingredient-parser
package.

As ingredient-parser has not yet been published to npm, we must refer to
it in package.json by the GitHub URL along with the git tag we'd like to
use.