Ignore leading and trailing whitespace when editing
Before, the pound sign indicating a section title needed to occur as the
first character on a line and be immediately followed by a space. Now,
we are more flexible about this: we ignore leading and trailing
whitespace on each line, and do not require a space to follow the pound
sign. The pound sign only needs to be the first non-space character on a
line in order to be treated as a title.