Do not require blank lines between instructions
After playing with the editor a bit more, I think this is closer to what
I'd expect as a user (pressing return leads to a new step). The
formatter which creates the editable text from the JSON continues to
insert blank lines between steps as this aids in comprehension, but
these blank lines can be removed with no impact if desired.