Refresh user page
Refactor the EditableImage component to do less. Now, it is no longer
responsible for actually displaying the image and only displays the Edit
button (which has also been simplified). Add a route to the API for
handling edits to the logged in user by making a PUT request to
/api/user. Change the user page to have a side-by-side layout of some
basic profile details on the left and a selection of recipes on the
right, with support for changing the avatar and display name when the
user is logged in and looking at their own page.