Move full recipe list onto user page
As the /username and /username/recipes pages were extremely similar, it
turns out we can quite nicely collapse them into a single page and
provide a better user experience. The RecipePreview component has been
removed in favor of simply using the recipe list component, and the
search for recipes has been moved from /username/recipes?q= to just
/username?q=.

This also means we can remove the "Recipes" link from the navbar, as it
would lead to a 404 page.