Make the recipe list items sized consistently
Re-add a fallback image to the recipe list for display when the user
hasn't uploaded an image for the recipe. The image is displayed with a
consistent size (80x50px, the golden ratio) to the left of each recipe
title.

Remove the "by {{username}}" fallback for cases where there is no
description. Since this list is now only ever displayed on the user's
profile page, it's much more obvious who the recipes belong to, and it
should also be less confusing as to why a recipe's description appears
to be "by {{username}}" (since it will no longer say that at all).