Update instrumentation
This patch adjusts our initial instrumentation to hopefully be a bit
more helpful. The Prometheus metrics have been moved from the API router
to be an app-level middleware so that we capture all HTTP requests.
Also, the metrics have been renamed to use the Prometheus-recommended
naming scheme.

A "handler" label has been added so that we can more easily identify
problematic or frequently used handler functions. This name is taken
from the actual function name which is used to handle the request, so
all of the API handler functions have been changed from anonymous
functions to named functions, which enables this tracking.

The importer now collects a few additional metrics as well: for website
imports, we have a counter labelled by hostname so we can see which
sites we are predominantly importing from. The file importer now tracks
number of files in each upload as well as the size of each file. These
metrics are labelled by MIME type, again so that we can see what sorts
of imports are most popular.