Implement notes
Add a new notes table to the database, along with accompanying models. A
Note represents loose commentary about a recipe, such as "next time, use
more parsley," or "this is tasty." A note is related to a particular
version of a recipe, though it is also related to a recipe directly so
in the future we may support notes which refer to a recipe more
generally and are not also linked to a specific version.  Currently, all
notes are public and only a recipe's owner may write notes about it.

A note may also be pinned; pinning a note means that it will be visible
at the top of the master recipe view page until it is un-pinned. This is
in order to support notes that don't go away when a new version of the
recipe is created, such as "Be sure to start rice!" or "Use pint sized
jars."

In order to handle CRUD operations for notes, new API routes have been
added, both under /api/user/notes (for create, update, delete) and under
/api/users/{username}/recipes/{slug}/notes (for read).

A new Notes tab appears in the RecipeNav bar and provides a place to
view and add notes about the recipe version. A toggle switch lets users
choose whether they want to see only notes on the current version, or
all notes about the recipe.

In order more clearly represent the recipe/version hierarchy, the recipe
navigation has been updated. To start, there is a new RecipeLayout
component which serves as a replacement for the Layout component on
pages that display recipe information. The version-specific items, such
as the yield and duration, have been moved into a new ribbon at the top
of the RecipeVersion component along with Preheats. Additionally, there
is a "condensed" prop on RecipeLayout which can be set to true in order
to show a slightly smaller header image on the ancillary pages (e.g.
History, Notes).

The RecipeNav bar is now more contextual. There are two general modes of
navigation, top-level recipe navigation, where the master version is
being browsed, and version-specific navigation where a particular
version is being examined. Visiting /{user}/{recipe} starts off in
top-level recipe navigation. The recipe is displayed with the
large-style header image (if available), along with its description and
pinned notes. By going to the History tab and selecting a previous
version, the user will be taken to /{user}/{recipe}/versions/{id} and
will enter the version-specific mode of navigation. Now, tapping the
History or Notes tab will bring them to
/{user}/{recipe}/versions/{id}/history or /notes where they will see
versions of those pages specific to the version they are currently
browsing. While in version-specific navigation, tapping the Recipe tab
displays only the recipe version they are browsing. This page shows a
warning explaining that they are viewing a specific version, the
"commit" message for it, and a link back to top-level recipe navigation.