Fix unit tests
Address failing tests. In the future, we should move testing for the
ingredients out into the ingredient-parser library and only test the
list functionality here in the importer tests.

Create a testAsset helper function which returns a promise for the
content of a file in the test/assets directory. We can use this instead
of readFileSync for speed.

Change the importer tests to only run the importer once on each test
file, then run several tests against the result. This way, we're not
wasting time running the (potentially somewhat expensive) import process
multiple times in a test suite. This also helps improve the organization
of the test files somewhat.