Create dropdown-item-danger class
Reactstrap provides a cssModule prop which hooks into the ability to
swap out CSS classes. In our case, we want to have a "danger"-styled
DropdownItem for actions like deleting something.

Add a dropdown-item-danger class which extends Bootstrap's built-in
dropdown-item class and sets the text color and hover background, and
use the cssModule prop to swap out the dropdown-item class for
dropdown-item-danger.