Change preparation parsing to info color
Before, we displayed the parsed preparation in the muted color which did
not provide enough contrast from the background. Now, we use the info
color which pops quite a bit more. Additionally, we now consistently
highlight the amount and preparation on the editor page, but show them
in black when viewing the recipe. This makes it look less like a mistake
to have the amount not be green on the view page.

Fixes #257
Fixes #179