Add full-text recipe search
Before, we were using a primitive Postgres ILIKE comparison on the
recipe title and description fields in order to figure out which
recipes were relevant to a search query. This worked decently well, but
it turns out that a common way people want to search their recipes is by
ingredient as well. Rather than implementing an "advanced" search
feature where you can go crazy and select which fields you want to
search for what terms and do a whole bunch of Boolean logic, this patch
adds support for Postgres' built-in full text search indexing.

For background on this method, read the excellent blog post at
http://rachbelaid.com/postgres-full-text-search-is-good-enough/.

There are a few parts to this. First, a view has been added to the
database called recipe_search_documents. This view is designed to
provide a clean interface for running queries and it takes care of
collating the procedures and ingredients of any recipe branch's current
version with the title and description and building a tsvector from
those words (tsvectors are what can be searched over). Effectively, this
searches the recipe's current master branch as that is the only branch
we currently support, though in the future when we expand out to having
multiple branches, the search would also cover other branches. This is
basically just a way to pull in the current version; if we were to
implement better branch support we'd likely want to update the search
functionality to return the specific branch of the recipe that matches,
but that is a problem for the future.

Sequelize doesn't support working with views, so we trick it into
thinking that recipe_search_documents is just a table so that we can do
all the fun sequelize things with it, like including the related recipe
model in the query results.

At the API layer, the ability to add a ?q= query parameter to the
/recipes endpoint has been removed (though they can still be sorted) and
all recipe searches use the previously-implemented-but-hereto-unused
/search endpoint.

The limited UI changes required mostly revolve around knowing when to
use the /recipes and when to use the /search endpoint in order to grab
the list of recipes. The debounce time has been upped from 200ms to
500ms in an effort to run the more-expensive full-text search only when
the user stops typing, and a new "Relevance" option has been added to
the sort order (which, in practice, simply sets the sort to '' as
ranking by relevance doesn't have an associated column and is the
default).

I fully expect we will need to tune a bunch of these things once this is
deployed. Potential performance improvements include using a
materialized view rather than a "regular" view for
recipe_search_documents, and updating the search page to only run the
query once the user hits enter versus searching as-you-type.

The reason this is not yet a materialized view is because that requires
an external refresh process, and would introduce a delay into the recipe
search. For example, if a user adds a recipe for "roasted turnips" and
then immediately searches for "turnip," their newly added recipe would,
confusingly, not appear in the search results until the next time the
materialized view was refreshed. Even refreshing frequently, e.g. every
minute, this might still result in confusing delays. Since the
performance seems to be good enough right now to simply run the full
text search without materializing and indexing the view, we can defer
solving this problem for a little while.