OCR: Append Text and Rotate Images (#296)
Some quick quality of life improvements for the OCR tool. All text is
now appended instead of replacing. Rotating the image is now supported,
but ugly as heck. But it works. Also dumps the last selected region to
the disk so you can select something easily to be used as an image.