Consolidate fetching of RecipeLayout props
The RecipeLayout component has been helpful for displaying all of the
recipe sub-pages with a consistent layout, but the way its props are
being passed in has left much to be desired. Before, every page needed
to fetch a whole bunch of very similar data to be passed into the
RecipeLayout component's props.

To improve this situation, add a fetchRecipeLayoutProps function which
does just what it says on the tin: given some of the context from the
static getInitialProps functions on the page components, return the
props needed to render the RecipeLayout. This does depend on using the
same URL parameter names, which so far we have been doing.

In order to keep track of notes being added/updated/removed, a new
RecipeContext has also been introduced. The RecipeLayout component wraps
its children in a RecipeContext.Provider, so all of the components
rendered on the recipe sub-pages have access to the shared context in a
consistent way.

In addition to providing details on the recipe/version, the
RecipeContext also provides a mechanism to add/update/delete notes in
the local state without reloading any data from the server. This works
by using the new useReducer hook from React, whose API mirrors that of
Redux (but of course does not require any additional dependencies). The
RecipeLayout gets an initial list of notes from the server when the page
is loaded, and these notes constitute the reducer's initial state. The
RecipeContext provides event hooks addNote, updateNote, and removeNote
which in turn call the reducer's dispatch function in order to update
the context.