Shopping List Schema
A shopping list has a name. It contains many shopping list items, which
also have names and completed statuses. A user may have many shopping
lists. When they create it themselves, they are the owner. Another user
may be added to an existing shopping list and they are marked as not an
owner. Only the owner will be able to remove a shopping list.