Add user activity tracking
Create a new user_activity table which records whether a user was active
on a given day. Each time an authenticated request is made to the API,
we try to insert a row for the user into the user_activity table. Since
we have a unique index on day and user_id, and the INSERT query uses "on
conflict do nothing," the result is simply a row in the table for each
day that a user was active on PlateZero. We can then use this number to
calculate our active user count, for example:

    SELECT count(*) FROM (
      SELECT DISTINCT user_id FROM user_activity
      WHERE date >= now() - '7 days'::interval
    ) AS active_user_ids;