Shopping Lists (#309)
Add a new page to PlateZero to display shopping lists for the current
logged in user. Only logged in users can view their shopping lists. A
shopping list has a name and contains items, which also have names and a
completed status.

Current status is creation of list and items and completing them. More
work will be done for editing and removing, as well as better handling
of reordering the items and hiding completed items.

More integration will occur between recipes and your shopping lists.