Restore the scroll position during history navigation
Because rendering the page in the browser with React fucks with the
browser's automatic scroll restoration, we can hack around this a bit by
saving the position they had scrolled to before each navigation and
hopping back there if it's a Back event (history state pop).

This isn't perfect, but should fix the specific annoyance of getting
scrolled back up to the top of a long list of recipes and losing your
place.

Closes #312