Split out UserPageRecipes
In preparation for adding multiple tabs on the user page, extract the
logic to show the searchable recipe list based on query parameters into
a separate component. This has the pleasant side effect of making the
logic much clearer, because previously it was included in a nextjs page
component which needs to be implemented as a class, and this is the sort
of state management where react hooks really shine.